that works because the instruction is a push imm32 not push imm8

+++ t4m/PatchT4Dvars.cpp
@@ -19,7 +19,7 @@ void PatchT4_Dvars()
 {
 	//DVAR: cg_fov
 	//MODS: Clear cheat flag, set saved flag, set maximum to 90
-	*(BYTE*)0x65EE46 ^= DVAR_FLAG_CHEAT | DVAR_FLAG_SAVED; // How do you imagine saving a short in a single byte? You need a hook to apply the SAVED flag...
+	*(WORD*)0x65EE46 ^= DVAR_FLAG_CHEAT | DVAR_FLAG_SAVED; // How do you imagine saving a short in a single byte? You need a hook to apply the SAVED flag...
 	*(float**)0x65EE31 = &cgFov90;
 	
 	//
